fixing deployment

+++ db/migrate/20170928072635_add_two_significant_figures_to_price.rb
@@ -1,5 +1,5 @@
 class AddTwoSignificantFiguresToPrice < ActiveRecord::Migration
   def change
-    change_column :products, :price, :decimal, precision: 5, scale: 2
+    change_column :products, :price, :float, precision: 5, scale: 2
   end
 end
